Refactor risk management utilities into reusable modules
