+++ risk_management/account_clients.py
@@ -4,7 +4,6 @@ from __future__ import annotations
 
 import abc
 import asyncio
-import json
 import logging
 import math
 import statistics
@@ -28,6 +27,18 @@ from custom_endpoint_overrides import (
 )
 
 from risk_management.configuration import AccountConfig
+from ._utils import (
+    coerce_float as _coerce_float,
+    coerce_int as _coerce_int,
+    extract_position_details as _extract_position_details,
+    first_float as _first_float,
+    stringify_payload as _stringify_payload,
+)
+from ._parsing import (
+    extract_balance as _extract_balance,
+    parse_order as _parse_order,
+    parse_position as _parse_position,
+)
 from .realized_pnl import fetch_realized_pnl_history
 
 try:  # pragma: no cover - passivbot is optional when running tests
@@ -40,138 +51,6 @@ except (ModuleNotFoundError, ImportError):  # pragma: no cover - allow running w
 
 
 logger = logging.getLogger(__name__)
-
-
-def _json_default(value: Any) -> Any:
-    """Coerce non-serialisable objects into JSON-compatible types."""
-
-    if isinstance(value, (set, frozenset, tuple)):
-        return list(value)
-    if isinstance(value, bytes):
-        try:
-            return value.decode("utf-8")
-        except UnicodeDecodeError:
-            return value.decode("utf-8", errors="replace")
-    return str(value)
-
-
-def _stringify_payload(payload: Any) -> str:
-    """Return a JSON string representation for logging purposes."""
-
-    try:
-        return json.dumps(payload, ensure_ascii=False, default=_json_default, sort_keys=True)
-    except (TypeError, ValueError):  # pragma: no cover - defensive fallback
-        return repr(payload)
-
-
-def _first_float(*values: Any) -> Optional[float]:
-    """Return the first value that can be coerced into ``float``."""
-
-    for value in values:
-        candidate = value
-        if candidate is None:
-            continue
-        if isinstance(candidate, (list, tuple)) and candidate:
-            candidate = candidate[0]
-        try:
-            return float(candidate)
-        except (TypeError, ValueError):
-            if isinstance(candidate, str):
-                try:
-                    return float(candidate.strip())
-                except (TypeError, ValueError):
-                    continue
-            continue
-    return None
-
-
-def _coerce_float(value: Any) -> Optional[float]:
-    """Return ``value`` converted to ``float`` when possible."""
-
-    if value is None:
-        return None
-    try:
-        return float(value)
-    except (TypeError, ValueError):
-        if isinstance(value, str):
-            try:
-                return float(value.strip())
-            except (TypeError, ValueError):
-                return None
-    return None
-
-
-def _coerce_int(value: Any) -> Optional[int]:
-    """Return ``value`` converted to ``int`` when possible."""
-
-    if value is None:
-        return None
-    if isinstance(value, bool):
-        return int(value)
-    try:
-        return int(value)
-    except (TypeError, ValueError):
-        if isinstance(value, str):
-            try:
-                return int(float(value.strip()))
-            except (TypeError, ValueError):
-                return None
-    return None
-
-
-def _normalize_position_side(value: Any) -> Optional[str]:
-    """Return the normalised hedge-mode side when available."""
-
-    if isinstance(value, str):
-        candidate = value.strip().upper()
-        if candidate in {"LONG", "SHORT", "BOTH"}:
-            return candidate
-    return None
-
-
-def _extract_position_details(
-    position: Mapping[str, Any]
-) -> Tuple[Optional[str], Optional[int], bool]:
-    """Return the detected hedge side, index, and whether the side was explicit."""
-
-    position_side = _normalize_position_side(position.get("positionSide"))
-    side_explicit = position_side is not None
-    if position_side is None:
-        alt_side = _normalize_position_side(position.get("position_side"))
-        if alt_side is not None:
-            position_side = alt_side
-            side_explicit = True
-
-    def _position_idx_from(obj: Any) -> Optional[int]:
-        if not isinstance(obj, Mapping):
-            return None
-        raw_idx = obj.get("positionIdx", obj.get("position_idx"))
-        if raw_idx is None:
-            return None
-        try:
-            value = int(float(raw_idx))
-        except (TypeError, ValueError):
-            return None
-        if value in {0, 1, 2}:
-            return value
-        return None
-
-    position_idx = _position_idx_from(position)
-    info = position.get("info")
-    if position_idx is None and isinstance(info, Mapping):
-        position_idx = _position_idx_from(info)
-
-    if position_side is None and isinstance(info, Mapping):
-        info_side = _normalize_position_side(info.get("positionSide") or info.get("position_side"))
-        if info_side is not None:
-            position_side = info_side
-            side_explicit = True
-
-    if position_side is None and position_idx in {1, 2}:
-        position_side = "LONG" if position_idx == 1 else "SHORT"
-
-    return position_side, position_idx, side_explicit
-
 
 class AccountClientProtocol(abc.ABC):
     """Abstract interface for realtime account clients."""
@@ -687,12 +566,6 @@ class CCXTAccountClient(AccountClientProtocol):
         await self._ensure_markets()
         balance_raw = await self.client.fetch_balance(params=self._balance_params)
         self._log_exchange_payload("fetch_balance", balance_raw, self._balance_params)
-        from .realtime import (  # circular safe import
-            _extract_balance,
-            _parse_order,
-            _parse_position,
-        )
-
         balance_value = _extract_balance(balance_raw, self.config.settle_currency)
         positions_raw: Iterable[Mapping[str, Any]] = []
         positions: list[Dict[str, Any]] = []
@@ -1037,8 +910,6 @@ class CCXTAccountClient(AccountClientProtocol):
             )
             raise RuntimeError(f"Exchange order placement failed: {exc}") from exc
         self._log_exchange_payload("create_order", response, {**request_params, **payload})
-        from .realtime import _parse_order  # circular safe import
-
         normalized = _parse_order(response) if response else None
         return {"order": normalized, "raw": response or {}}
 
@@ -1068,8 +939,6 @@ class CCXTAccountClient(AccountClientProtocol):
             raise RuntimeError(f"Exchange order cancellation failed: {exc}") from exc
         payload_meta = {"order_id": order_id, "symbol": symbol}
         self._log_exchange_payload("cancel_order", response, {**request_params, **payload_meta})
-        from .realtime import _parse_order  # circular safe import
-
         normalized = _parse_order(response) if response else None
         return {"order": normalized, "raw": response or {}}
